Data-ETL-Pipeline/requirements.txt
requests==2.31.0
selenium==4.16.0
webdriver-manager==4.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
flask==3.0.0
pytest==7.4.3
gunicorn==21.2.0
statsbombpy==1.13.0
beautifulsoup4==4.12.2
lxml==4.9.3
apscheduler==3.10.4
